code/envsense/requirements.txt
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
colorzero==1.1
coverage==5.3
gpiozero==1.5.1
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
six==1.15.0
smbus2==0.3.0
toml==0.10.2
wrapt==1.12.1